Refactor TagManager into modular components with improved code organization
